Expose test utils for engine and mvcc (#3444)
